backend/requirements.txt
requests
Pillow==10.0.0
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
mysql-connector-python==8.2.0
python-dotenv==1.0.0
flask-mail==0.9.1
flask_migrate==4.0.5